refactor(#806): extract UpdateExpression codegen → unary-updates.ts
Split src/codegen/expressions/unary.ts (1643 LOC mixing concerns) into:

- unary.ts — 151 LOC, non-update prefix unary (+, -, !, ~).
- unary-updates.ts — 1537 LOC, all UpdateExpression codegen
  (unwrapParens, emitToNumericForUpdate, compileMemberIncDec,
   compilePrefixUpdate, compilePostfixUnary, compile{Prefix,Postfix}
   Increment{Property,Element}).

compilePrefixUnary now delegates the PlusPlusToken / MinusMinusToken
switch arms to a new compilePrefixUpdate helper that lives alongside
the rest of the update logic. No behavioural change — public exports
from ./unary.js are preserved by re-export.

Move the issue from backlog/ → sprints/53/ and mark status: done.
Re-scoped from the original "extract from expressions.ts" goal (already
completed in a prior split) to a finer separation inside expressions/.

Validation:
- pnpm run build green
- 14/14 scoped smoke tests (locals, props, elements, captures)
- 32/32 equivalence tests (prefix-postfix-increment-property,
  compound-assignment-*, unary-plus-coercion-*)
